выделяет варианты сортировки в отдельные компоненты
